refactor(types): rename control panel and menu type files and update imports
